[#31] feat : 유저와 연관관계 설정 및 관련 검증 로직 각 api에 추가

+++ src/main/java/com/b6/mypaldotrip/domain/review/service/ReviewService.java
@@ -13,7 +13,9 @@ import com.b6.mypaldotrip.domain.review.store.repository.ReviewRepository;
 import com.b6.mypaldotrip.domain.trip.service.TripService;
 import com.b6.mypaldotrip.domain.trip.store.entity.TripEntity;
 import com.b6.mypaldotrip.global.exception.GlobalException;
+import com.b6.mypaldotrip.global.security.UserDetailsImpl;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,13 +27,19 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
     private final TripService tripService;
 
-    public ReviewCreateRes createReview(Long tripId, ReviewCreateReq req) {
+    public ReviewCreateRes createReview(
+            Long tripId, ReviewCreateReq req, UserDetailsImpl userDetails) {
         TripEntity trip = tripService.findTrip(tripId);
         ReviewEntity review =
-                ReviewEntity.builder().content(req.content()).score(req.score()).trip(trip).build();
+                ReviewEntity.builder()
+                        .user(userDetails.getUserEntity())
+                        .content(req.content())
+                        .score(req.score())
+                        .trip(trip)
+                        .build();
         reviewRepository.save(review);
-        // TODO: 2024-01-11 response에 유저 정보 추가 필요
         return ReviewCreateRes.builder()
+                .username(userDetails.getUsername())
                 .content(review.getContent())
                 .score(review.getScore())
                 .modifiedAt(review.getModifiedAt())
@@ -45,6 +53,7 @@ public class ReviewService {
                 .map(
                         review ->
                                 ReviewListRes.builder()
+                                        .username(review.getUser().getUsername())
                                         .content(review.getContent())
                                         .score(review.getScore())
                                         .modifiedAt(review.getModifiedAt())
@@ -53,24 +62,38 @@ public class ReviewService {
     }
 
     @Transactional
-    public ReviewUpdateRes updateReview(Long tripId, Long reviewId, ReviewUpdateReq req) {
+    public ReviewUpdateRes updateReview(
+            Long tripId, Long reviewId, ReviewUpdateReq req, UserDetailsImpl userDetails) {
         TripEntity trip = tripService.findTrip(tripId);
         ReviewEntity review = findReview(reviewId);
+
         matchReviewAndTrip(review, trip);
+        checkAuthor(userDetails, review);
         review.updateReview(req.content(), req.score());
+
         return ReviewUpdateRes.builder()
+                .username(review.getUser().getUsername())
                 .content(review.getContent())
                 .score(review.getScore())
                 .modifiedAt(review.getModifiedAt())
                 .build();
     }
 
-    public ReviewDeleteRes deleteReview(Long tripId, Long reviewId) {
+    public ReviewDeleteRes deleteReview(Long tripId, Long reviewId, UserDetailsImpl userDetails) {
         TripEntity trip = tripService.findTrip(tripId);
         ReviewEntity review = findReview(reviewId);
         matchReviewAndTrip(review, trip);
+        checkAuthor(userDetails, review);
         reviewRepository.delete(review);
         return ReviewDeleteRes.builder().message("리뷰가 삭제되었습니다.").build();
+    }
+
+    // 작성자 검증
+    private static void checkAuthor(UserDetailsImpl userDetails, ReviewEntity review) {
+        if (!Objects.equals(
+                review.getUser().getUserId(), userDetails.getUserEntity().getUserId())) {
+            throw new GlobalException(ReviewErrorCode.NOT_AUTHOR_ERROR);
+        }
     }
 
     // 해당 리뷰가 위치하는 여행정보가 맞는지 검증
